Remove `quesma/query_processor` (move to `quesma/model`)
