+++ quesma/model/pipeline_aggregations/min_bucket.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"quesma/logger"
 	"quesma/model"
-	"quesma/queryprocessor"
 	"quesma/util"
 )
 
@@ -43,7 +42,7 @@ func (query MinBucket) CalculateResultWhenMissing(parentRows []model.QueryResult
 	if len(parentRows) == 0 {
 		return resultRows // maybe null?
 	}
-	qp := queryprocessor.NewQueryProcessor(query.ctx)
+	qp := model.NewQueryProcessor(query.ctx)
 	parentFieldsCnt := len(parentRows[0].Cols) - 2 // -2, because row is [parent_cols..., current_key, current_value]
 	// in calculateSingleAvgBucket we calculate avg all current_keys with the same parent_cols
 	// so we need to split into buckets based on parent_cols
